refactor(ir): keep call targets as templates
