+++ docs/spec/hir.md
@@ -115,39 +115,33 @@ declares, but it MUST NOT create a level or change one's extent.
 `return_type`. The projection is fixed at construction and stays
 consistent across construction sites.
 
-**Call typing — elaboration.** The template a `@func` declares lives at the
-Python-source level; IR never carries a template object or a shared
-polymorphic body. A `Call` whose target is a `Function` types by
-*elaboration*: `elaborate(callee, arg_types)` binds each parameter to the
-caller's actual argument type, then reconstructs the body under that
-binding — every node the reconstruction touches is typeinferred afresh and
-stamped exactly once, so **the callee specializes per call site** and a
-caller-supplied layout (sharding) flowing into a layout-unconstrained
-parameter propagates through the whole body, including through a `Tuple` or
-`GridRegionExpr` return. The result of elaboration is the concrete
-`Function` instance that becomes the `Call`'s `target`; the `Call`'s type is
-that instance's (re-derived) body type, never a stale `return_type` field
-carried over from a different call site.
+**Call typing — feed-driven inference.** A `Call` keeps its authored
+`Function` template as `target`. Its result type is inferred by feeding the
+actual argument types to the callee's formal parameters and walking the
+callee body in a child context. This is type inference only: it does not
+rebuild a `Function`, mutate the target, or create a per-call instance.
+Caller-supplied layout (sharding) flowing into a layout-unconstrained
+parameter propagates through the body, including through a `Tuple` or
+`GridRegionExpr` return.
 
 How a `Call` binds is part of what it states: which declared parameters
 `Call.args` supply, and the scope
 ([visitor-registry §4](./visitor-registry.md#4-instance-1--typeinfer)) the
 callee's body is read in. Argument types bind to the supplied parameters in
 order; a parameter the call does not supply keeps its declared type in the
-rebuilt signature, which for a `ConstTensor` is already concrete — no value it
-stands for enters the IR, and what fills it comes from outside
+feed, which for a `ConstTensor` is already concrete — no value it stands for
+enters the IR, and what fills it comes from outside
 ([runtime §1.1.2](./runtime.md#112-weight-converter-and-prepare--forward)). The
-binding is part of the elaboration cache key. It MUST be stated by the call in
-the scope it is read in, never counted from how many arguments the call passes,
-and it is not a question the context answers.
+binding MUST be stated by the call in the scope it is read in, never counted
+from how many arguments the call passes, and it is not a question the context
+answers.
 
 Omitting a parameter is valid only where, within that scope, the callee is
 uniquely owned by a direct child of the caller's owner
 ([core-ir §1](./core-ir.md#1-module)); before collection the authored binding
 answers the same question of the child it names. Ownership that is missing,
 ambiguous, or not a direct child supplies every declared parameter, as does
-`elaborate` with no call in hand, so a standalone or low-level call cannot
-acquire implicit constants.
+so a standalone or low-level call cannot acquire implicit constants.
 
 Caller and callee MUST resolve one effective `Target` and one effective topology
 hierarchy: a same-kernel call is one execution context, whichever `Module` owns
@@ -181,20 +175,11 @@ op, not at the boundary. A dispatch-prototype callee
 result is the declared `return_type` and the `None` body is never inspected
 (variant selection is **Shape dispatch and specializations** below).
 
-An elaborated instance MUST record the function it was rebuilt from, the same
-record a specialization writes (`origin_of`, **Function specialization API**
-below), so the instance in a `Call.target` stays connected to the function a
-Module owns without matching on the name they share. It records no bound
-dimensions: a call site binds parameter types, it does not choose an extent.
-
-Elaboration memoizes per construction session — one parser run, or one
-top-level `elaborate` call and every nested call it re-elaborates — keyed
-on (callee, argument types): two call sites of the same callee with
-identical argument types MUST resolve to the identical `Function`
-instance, not merely an equal one, so a viewer/printer keyed on instance
-identity renders one node per distinct specialization. The memo is
-session-local state, not module or process state; it carries nothing
-across sessions.
+Call typing does not record provenance or bound dimensions because no
+call-site function instance is created. Explicit specialization remains the
+separate operation that produces a derived `Function`; such a derived
+function records its origin and chosen dimensions as specified by the
+**Function specialization API** below.
 
 **Signature annotation `Layout.strides` materialization.** A
 `Tensor[..., (sugar)]` annotation on a parameter or return appears
